Initial v2 plumbing
