+++ ansible/plugins/modules/vault_load_secrets.py
@@ -50,7 +50,6 @@ files.region2:
   testbar: ~/ca.crt
 """
 
-import base64
 import os
 import subprocess
 import time
@@ -58,6 +57,9 @@ from collections.abc import MutableMapping
 
 import yaml
 from ansible.module_utils.basic import AnsibleModule
+
+from .load_secrets_v1 import sanitize_values, get_secrets_vault_paths
+
 
 ANSIBLE_METADATA = {
     "metadata_version": "1.1",
@@ -216,22 +218,10 @@ def flatten(dictionary, parent_key=False, separator="."):
     return dict(items)
 
 
-def sanitize_values(module, syaml):
+def sanitize_values_v2(module, syaml):
     """
-    Sanitizes the secrets YAML object. If a specific secret key has
-    s3.accessKey and s3.secretKey but not s3Secret, the latter will be
-    generated as the base64 encoding of both s3.accessKey and s3.secretKey.
-
-      secrets:
-        test:
-          s3.accessKey: "1234"
-          s3.secretKey: "4321"
-
-    will push three secrets at 'secret/hub/test':
-
-    s3.accessKey: 1234
-    s3.secretKey: 4321
-    s3Secret: czMuYWNjZXNzS2V5OiAxMjM0CnMzLnNlY3JldEtleTogNDMyMQ==
+    Sanitizes the secrets YAML object version 2.0
+    ..TODO..
 
     Parameters:
         module(AnsibleModule): The current AnsibleModule being used
@@ -241,104 +231,13 @@ def sanitize_values(module, syaml):
     Returns:
         syaml(obj): The parsed yaml object sanitized
     """
-    if not ("secrets" in syaml or "files" in syaml):
-        module.fail_json(
-            f"Values secrets file does not contain 'secrets' or"
-            f"'files' keys: {syaml}"
-        )
+    version = get_version(syaml)
+    if version != "2.0":
+        module.fail_json(f"Version expected is 2.0 but got: {version}")
 
-    secrets = syaml.get("secrets", {})
-    # We need to explicitely check for None because the file might contain the
-    # top-level 'secrets:' or 'files:' key but have nothing else under it which will
-    # return None and not {}
-    if secrets is None:
-        secrets = {}
-    files = syaml.get("files", {})
-    if files is None:
-        files = {}
-    if len(secrets) == 0 and len(files) == 0:
-        module.fail_json(
-            f"Neither 'secrets' nor 'files have any secrets to be parsed: {syaml}"
-        )
-
-    if isinstance(secrets, list) or isinstance(files, list):
-        module.fail_json(f"Neither 'secrets' nor 'files can be lists: {syaml}")
-
-    for secret in secrets:
-        if not isinstance(secrets[secret], dict):
-            module.fail_json(
-                f"Each key under 'secrets' needs to point to "
-                f"a dictionary of key value pairs: {syaml}"
-            )
-
-    for file in files:
-        path = files[file]
-        if not os.path.isfile(os.path.expanduser(path)):
-            module.fail_json(f"File {path} does not exist")
-
-    # If s3Secret key does not exist but s3.accessKey and s3.secretKey do exist
-    # generate s3Secret so a user does not need to do it manually which tends to be error-prone
-    for secret in secrets:
-        tmp = secrets[secret]
-        if "s3.accessKey" in tmp and "s3.secretKey" in tmp and "s3Secret" not in tmp:
-            s3secret = (
-                f"s3.accessKey: {tmp['s3.accessKey']}\n"
-                f"s3.secretKey: {tmp['s3.secretKey']}"
-            )
-            s3secretb64 = base64.b64encode(s3secret.encode())
-            secrets[secret]["s3Secret"] = s3secretb64.decode("utf-8")
 
     return syaml
 
-
-def get_secrets_vault_paths(module, syaml, keyname):
-    """
-    Walks a secrets yaml object to look for all top-level keys that start with
-    'keyname' and returns a list of tuples [(keyname1, path1), (keyname2, path2)...]
-    where the path is the relative vault path
-    For example, given a yaml with the following:
-        secrets:
-            foo: bar
-        secrets.region1:
-            foo: baz
-        secrets.region2:
-            foo: barbaz
-
-    a call with keyname set to 'secrets' will return the following:
-    [('secrets', 'hub'), ('secrets', 'region1'), ('secrets', 'region2')]
-
-    Parameters:
-        module(AnsibleModule): The current AnsibleModule being used
-
-        syaml(obj): The parsed yaml object representing the secrets
-
-        keyname(str): The keytypes to look for either usually 'secrets' or 'files'
-
-    Returns:
-        keys_paths(list): List of tuples containing (keyname, relative-vault-path)
-    """
-    all_keys = syaml.keys()
-    keys_paths = []
-    for key in all_keys:
-        # We skip any key that does not start with 'secrets' or 'files'
-        # (We should probably bail out in the presence of unexpected top-level keys)
-        if not key.startswith(keyname):
-            continue
-
-        # If there is no '.' after secrets or files, assume the secrets need to
-        # go to the hub vault path
-        if key == keyname:
-            keys_paths.append((key, "hub"))
-            continue
-
-        # We are in the presence of either 'secrets.region-one' or 'files.cluster1' top-level keys
-        tmp = key.split(".", 1)
-        if len(tmp) != 2:
-            module.fail_json(f"values-secrets.yaml key is non-conformant: {key}")
-
-        keys_paths.append((key, tmp[1]))
-
-    return keys_paths
 
 
 # NOTE(bandini): we shell out to oc exec it because of
@@ -441,9 +340,20 @@ def run(module):
     syaml = parse_values(values_secrets)
     version = get_version(syaml)
 
-    if version != "1.0":
+    if version not in ["1.0", "2.0"]:
         module.fail_json(f"Version {version} is currently not supported")
 
+    if version == "2.0":
+        # In the future we can use the version field to manage different formats if needed
+        secrets = sanitize_values_v2(module, syaml)
+
+        #nr_secrets = inject_secrets(module, secrets, namespace, pod, basepath)
+        results["failed"] = False
+        results["changed"] = True
+        results["msg"] = f"{nr_secrets} secrets injected"
+        module.exit_json(**results)
+
+    # Default version 1.0
     # In the future we can use the version field to manage different formats if needed
     secrets = sanitize_values(module, syaml)
