+++ ansible/roles/vault_utils/tasks/push_secrets.yaml
@@ -39,7 +39,7 @@
     secret_template: "{{ pattern_dir }}/values-secret.yaml.template"
 
 - name: Loads secrets file into the vault of a cluster
-  no_log: true
+  no_log: false
   vault_load_secrets:
     values_secrets: ~/values-secret.yaml
     check_missing_secrets: false
