+++ ansible/tests/unit/v2/values-secret-v2-base.yaml
@@ -1,0 +1,45 @@
+# Needed to specify the new format (missing version means old version: 1.0 by default)
+version: 2.0
+# NEVER COMMIT THESE VALUES TO GIT
+
+# global + backend is the default. No need to specify it but one day we could have an override implementing it for, say, cyberark
+global:
+  backend: vault
+
+vaultPolicies:
+  # https://developer.hashicorp.com/vault/docs/concepts/password-policies
+  basicPolicy: |
+    length=10
+    rule "charset" { charset = "abcdefghijklmnopqrstuvwxyz" min-chars = 1 }
+    rule "charset" { charset = "ABCDEFGHIJKLMNOPQRSTUVWXYZ" min-chars = 1 }
+    rule "charset" { charset = "0123456789" min-chars = 1 }
+
+  advancedPolicy: |
+    length=20
+    rule "charset" { charset = "abcdefghijklmnopqrstuvwxyz" min-chars = 1 }
+    rule "charset" { charset = "ABCDEFGHIJKLMNOPQRSTUVWXYZ" min-chars = 1 }
+    rule "charset" { charset = "0123456789" min-chars = 1 }
+    rule "charset" { charset = "!@#$%^&*" min-chars = 1 }
+
+secrets:
+  - name: config-demo
+    vaultPrefixes:
+    - secret/region-one
+    - secret/snowflake.blueprints.rhecoeng.com
+    fields:
+    - name: secret
+      value: null
+      onMissingValue: error # One of: error,generate,prompt
+      vaultPolicy: basicPolicy
+    - name: secret2
+      value: null
+      onMissingValue: prompt
+  - name: config-demo-file
+    vaultPrefixes:
+    - secret/region-one
+    - secret/snowflake.blueprints.rhecoeng.com
+
+    file:
+    # promptable would prompt for the path of the file in this case
+      path: /tmp/ca.crt
+      onMissingValue: error # One of error, prompt (for path). generate makes no sense for file
